Refactor tests, remove extra logic in LoginForm
