+++ src/LoginForm/LoginForm.js
@@ -8,7 +8,7 @@ class LoginForm extends Component {
   }
 
   handleChange = (event) => {
-    let key = event.target.id || event.currentTarget.value;
+    let key = event.target.id;
     this.props.change(key, event);
   }
 
@@ -19,11 +19,11 @@ class LoginForm extends Component {
           <img className='logo' src={map} alt='VRAD map logo'/>
           <h1 className='VRAD-header-text'>VRAD</h1>
         </header>
-        <form className='login-form' onSubmit={this.props.loginUser}>
+        <form className='login-form' onSubmit={this.props.loginUser} aria-label='login'>
           <div className='error-container'></div>
-          <input type='text' id='name' placeholder='NAME' onChange={this.handleChange}/>
-          <input type='text' id='email' placeholder='EMAIL'onChange={this.handleChange}/>
-          <select id='type' onChange={this.handleChange} defaultValue='TYPE OF TRAVEL'>
+          <input type='text' id='name' value={this.props.formData.name} placeholder='NAME' onChange={this.handleChange}/>
+          <input type='text' id='email' value={this.props.formData.email} placeholder='EMAIL'onChange={this.handleChange}/>
+          <select id='type' value={this.props.formData.type} onChange={this.handleChange}>
             <option value='TYPE OF TRAVEL' disabled>TYPE OF TRAVEL </option>
             <option value='business' id='type'>BUSINESS</option>
             <option value='vacation'>VACATION</option>
