+++ src/LoginForm/LoginForm.test.js
@@ -9,36 +9,11 @@ import { Link, Route, BrowserRouter as Router } from 'react-router-dom';
 
 describe('LoginForm', () => {
   it('Renders with a header and form', () => {
-    const { getByText } = render(<LoginForm />);
+    const { getByText } = render(<LoginForm change={jest.fn()} loginUser={jest.fn()} formData={{name: '', email: '', type: 'TYPE OF TRAVEL'}}/>);
     const header = getByText('VRAD');
     const form = getByText('TYPE OF TRAVEL');
     expect(header).toBeInTheDocument();
     expect(form).toBeInTheDocument();
-  });
-
-  it('We should be able to login when form is complete', () => {
-    let mockLogin = jest.fn();
-    let mockChange = jest.fn();
-
-    const { getByPlaceholderText, getByText } = render(<Router> <LoginForm change={mockChange} loginUser={mockLogin}/> </Router>);
-
-    fireEvent.change(getByPlaceholderText('NAME'), {target: {value: 'Beyonce'}});
-    fireEvent.change(getByPlaceholderText('EMAIL'), {target: {value: 'Beyonce@gmail.com'}});
-    fireEvent.change(getByText('TYPE OF TRAVEL'), {target: {value: 'OTHER'}});
-    fireEvent.click(getByText('SUBMIT'));
-
-    expect(mockLogin).toHaveBeenCalledTimes(1);
-  });
-
-  it('We should not be able to login when form is incomplete', () => {
-    let mockLogin = jest.fn();
-
-    const { getByPlaceholderText, getByText } = render(<Router> <App loginUser={mockLogin}/> </Router>);
-
-    fireEvent.click(getByText('SUBMIT'));
-    let errorMessage = getByText('Please complete all required fields!')
-
-    expect(errorMessage).toBeInTheDocument();
   });
